+++ test/02_account_service.js
@@ -15,42 +15,38 @@ describe('resource account', () => {
     })
 
     it('list()', async () => {
-        let res = await account.list({
-            q: undefined,
+        let items = await account.list({
             p: 1
         })
 
-        assert(res.length > 0)
+        assert(items instanceof Array)
+        assert(items.length > 0)
     })
 
     it('create()', async () => {
-        let res = await account.create({
+        await account.create({
             email: new_email,
             password: created_password,
             role:'r'
         })
-        assert(res)
     })
 
     it('change_password()', async () => {
-        let res = await account.change_password({
+        await account.change_password({
             email: new_email,
             old_password: created_password,
             new_password: 'new password'
         })
-        assert(res)
     })
 
     it('change_role()', async () => {
-        let res = await account.change_role({
+        await account.change_role({
             email: new_email,
             role: 'rw'
         })
-        assert(res)
     })
 
     it('remove()', async () => {
-        let res = await account.remove(new_email)
-        assert(res)
+        await account.remove(new_email)
     })
 })